Fixed bug in pps-client-stop.cpp.

+++ utils/stop/pps-client-stop.cpp
@@ -12,7 +12,7 @@
 #include <fcntl.h>
 
 int main(void){
-	char cmd[100];
+	char cmd[500];
 
 	uid_t uid = geteuid();
 	if (uid != 0){
@@ -28,23 +28,31 @@ int main(void){
 	int fd = open("/run/shm/pps-client-pid.txt", O_RDONLY);
 
 	strcpy(cmd, "kill ");
-											// Append the PID value if any to cmd
-	read(fd, cmd + strlen(cmd), 100 - strlen(cmd));
+												// Append the PID value if any to cmd
+	read(fd, cmd + strlen(cmd), 500 - strlen(cmd));
 
-	char *pid = strpbrk(cmd, "0123456789");	// Locate the PID value
+	char *end = strchr(cmd, '\n');				// Get only the first line
+	if (end != NULL){
+		*end = '\0';							// by ending at '\n'.
+	}
+
+	if (strstr(cmd, "pps-client-") != NULL){	// This is not pps-client.
+		printf("pps-client is not running.\n");	// It's pps-client-stop or
+		return 0;								// pps-client-remove.
+	}
+
+	char *pid = strpbrk(cmd, "0123456789");		// Locate the PID value
 	if (pid == NULL){
-		printf("pps-client is closed.\n");
+		printf("pps-client is not running.\n");
 		return 0;
 	}
 
-	char *end = strpbrk(pid, " \r\n");		// Get the space after the PID value
-	*end = '\0';							// Terminate PID value at the space
+	end = strpbrk(pid, " \r\n");				// Get the space after the PID value
+	*end = '\0';								// Terminate PID value at the space
 
-	printf("Issuing the kill command:\n");
-	printf(cmd);
-	printf("\n");
+	printf("Closing pps-client\n");
 
-	system(cmd);							// Issue: "kill PID"
+	system(cmd);								// Issue: "kill PID"
 
 	return 0;
 }
